quick fix

+++ install.sh
@@ -18,7 +18,7 @@ git submodule update
 
 # Link dotfiles
 for f in bash_history bash_completion bash_profile bashrc gitconfig htoprc jshintrc screenrc tmux.conf vimrc vim; do
-	rm -r ~/."$f"
+	[[ -d ~/.$f ]] && rm -r ~/."$f"
 	symlink "$PWD/$f" ~/."$f"
 done
